+++ Level_4/miniRT/src/utils/to.c
@@ -1,6 +1,6 @@
 #include "minirt.h"
 
-int	to_decimal(char *s, float *f)
+int	to_decimal(char *s, double *f)
 {
 	//!TO BE ADDED OVERFLOWS CHECKS 
 	//TODO overflows, specific error such as 1. as a data etc
